Move purge worker into repo-updater and simplify init
