+++ internal/batches/syncer/types.go
@@ -4,9 +4,12 @@ import (
 	"context"
 
 	"github.com/sourcegraph/sourcegraph/internal/api"
+	"github.com/sourcegraph/sourcegraph/internal/goroutine"
 )
 
 type ChangesetSyncRegistry interface {
+	goroutine.BackgroundRoutine
+
 	UnarchivedChangesetSyncRegistry
 	// EnqueueChangesetSyncs will queue the supplied changesets to sync ASAP.
 	EnqueueChangesetSyncs(ctx context.Context, ids []int64) error
