+++ internal/repos/metrics.go
@@ -38,14 +38,4 @@ var (
 		Name: "src_repoupdater_syncer_synced_repos_total",
 		Help: "Total number of synced repositories",
 	}, []string{tagState})
-
-	purgeSuccess = promauto.NewCounter(prometheus.CounterOpts{
-		Name: "src_repoupdater_purge_success",
-		Help: "Incremented each time we remove a repository clone.",
-	})
-
-	purgeFailed = promauto.NewCounter(prometheus.CounterOpts{
-		Name: "src_repoupdater_purge_failed",
-		Help: "Incremented each time we try and fail to remove a repository clone.",
-	})
 )
